pyproject.toml
[tool.poetry]
name = "discord-music-player-bot"
version = "1.0.0"
description = ""
authors = ["kiryanchi <kiryanchi@naver.com>"]
readme = "README.md"
packages = [{include = "discord_music_player_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
youtube-search-python = "^1.6.6"
sqlalchemy = "^2.0.4"
rich = "^13.3.1"
yt-dlp = "^2023.3.4"
discord-py = {extras = ["voice"], version = "^2.2.2"}
wavelink = "^2.1.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
